Organizing service calls: Move ngwmn service calls into services package.
