+++ server/ngwmn/tests/services/test_ngwmn.py
@@ -6,7 +6,7 @@ from unittest import TestCase, mock
 
 import requests as r
 
-from ..fetch_utils import generate_bounding_box_values, get_well_lithography
+from ngwmn.services.ngwmn import generate_bounding_box_values, get_well_lithography
 
 
 class TestGetWellLithography(TestCase):
@@ -17,7 +17,7 @@ class TestGetWellLithography(TestCase):
         self.test_location_id = 'BP-1729'
         self.test_xml = '<site><agency>DOOP</agency><id>BP-1729</id></site>'
 
-    @mock.patch('ngwmn.fetch_utils.r.get')
+    @mock.patch('ngwmn.services.ngwmn.r.get')
     def test_success(self, r_mock):
         m_resp = mock.Mock(r.Response)
         m_resp.content = self.test_xml
@@ -30,7 +30,7 @@ class TestGetWellLithography(TestCase):
             params={'request': 'well_log', 'agency_cd': 'DOOP', 'siteNo': 'BP-1729'}
         )
 
-    @mock.patch('ngwmn.fetch_utils.r.get')
+    @mock.patch('ngwmn.services.ngwmn.r.get')
     def test_service_failure(self, r_mock):
         m_resp = mock.Mock(r.Response)
         m_resp.status_code = 500
@@ -38,7 +38,7 @@ class TestGetWellLithography(TestCase):
         result = get_well_lithography(self.test_service_root, self.test_agency_cd, self.test_location_id)
         self.assertIsNone(result)
 
-    @mock.patch('ngwmn.fetch_utils.r.get')
+    @mock.patch('ngwmn.services.ngwmn.r.get')
     def test_syntax_error(self, r_mock):
         m_resp = mock.Mock(r.Response)
         m_resp.content = 'Stuff'
